requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
